Wrap new logic in module
